GSYE-327: Moved offer_factory and bid_factory methods to the __init__ module of the test_matching_algorithms package, in order to be reused by multiple test classes.
