+++ tests/test_matching_algorithms/test_preferred_partners_matching_algorithm.py
@@ -15,57 +15,24 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
-# pylint: disable=missing-function-docstring, protected-access
-import uuid
-from typing import Optional
+# pylint: disable=protected-access
 
-from pendulum import DateTime
-
-from gsy_framework.data_classes import Offer, BidOfferMatch, Bid
-from gsy_framework.matching_algorithms.preferred_partners_algorithm import \
-    PreferredPartnersMatchingAlgorithm
+from gsy_framework.data_classes import BidOfferMatch
+from gsy_framework.matching_algorithms.preferred_partners_algorithm import (
+    PreferredPartnersMatchingAlgorithm)
+from tests.test_matching_algorithms import offer_factory, bid_factory
 
 
 class TestPreferredPartnersMatchingAlgorithm:
     """Tester class for the PreferredPartnersMatchingAlgorithm."""
-    @staticmethod
-    def offer_factory(additional_data: Optional[dict] = None):
-        """Create and return an offer object from default and input values."""
-        additional_data = additional_data or {}
-        return Offer(
-            **{"id": str(uuid.uuid4()),
-               "creation_time": DateTime.now(),
-               "time_slot": DateTime.now(),
-               "price": 10,
-               "energy": 30,
-               "seller": "seller",
-               "seller_id": "seller_id",
-               "seller_origin": "seller",
-               "seller_origin_id": "seller_id",
-               **additional_data})
 
     @staticmethod
-    def bid_factory(additional_data: Optional[dict] = None):
-        """Create and return a bid object from default and input values."""
-        additional_data = additional_data or {}
-        return Bid(
-            **{"id": str(uuid.uuid4()),
-               "creation_time": DateTime.now(),
-               "time_slot": DateTime.now(),
-               "price": 10,
-               "energy": 30,
-               "buyer": "buyer",
-               "buyer_id": "buyer_id",
-               "buyer_origin": "buyer",
-               "buyer_origin_id": "buyer_id",
-               **additional_data})
-
-    def test_get_matches_recommendations(self):
+    def test_get_matches_recommendations():
         """Test the main interface of the algorithm.
          Pass supported format data and receive correct results
          """
-        offer = self.offer_factory().serializable_dict()
-        bid = self.bid_factory(
+        offer = offer_factory().serializable_dict()
+        bid = bid_factory(
             {"requirements": [{"trading_partners": [offer["seller_id"]]}]}
         ).serializable_dict()
         data = {"market": {"2021-10-06T12:00": {
@@ -83,8 +50,9 @@ class TestPreferredPartnersMatchingAlgorithm:
                                      "offer_requirement": {}
                                  }).serializable_dict()]
 
-    def test_get_energy_and_clearing_rate(self):
-        offer = self.offer_factory().serializable_dict()
+    @staticmethod
+    def test_get_energy_and_clearing_rate():
+        offer = offer_factory().serializable_dict()
         assert PreferredPartnersMatchingAlgorithm._get_required_energy_and_rate_from_order(
             order=offer, order_requirement={}
         ) == (offer["energy"], offer["energy_rate"])
@@ -94,15 +62,17 @@ class TestPreferredPartnersMatchingAlgorithm:
             order=offer, order_requirement=order_requirement) == (
             order_requirement["energy"], order_requirement["price"] / order_requirement["energy"])
 
-    def test_get_actors_mapping(self):
+    @staticmethod
+    def test_get_actors_mapping():
         offers = [
-            self.offer_factory({"id": f"id-{index}",
-                                "seller_id": f"seller_id-{index}",
-                                "seller": f"seller-{index}",
-                                "seller_origin_id": f"seller_id-{index}",
-                                "seller_origin": f"seller-{index}"}).serializable_dict()
+            offer_factory({
+                "id": f"id-{index}",
+                "seller_id": f"seller_id-{index}",
+                "seller": f"seller-{index}",
+                "seller_origin_id": f"seller_id-{index}",
+                "seller_origin": f"seller-{index}"}).serializable_dict()
             for index in range(3)]
-        offers.append(self.offer_factory({
+        offers.append(offer_factory({
             "seller_id": offers[0]["seller_id"],
             "seller_origin_id": "different_origin_id",
             "seller_origin": "different_origin"}).serializable_dict())
@@ -112,10 +82,11 @@ class TestPreferredPartnersMatchingAlgorithm:
             "seller_id-2": [offers[2]],
             "different_origin_id": [offers[3]]}
 
-    def test_can_order_be_matched(self):
-        bid = self.bid_factory(
+    @staticmethod
+    def test_can_order_be_matched():
+        bid = bid_factory(
             {"requirements": [{"energy_type": ["green"]}]}).serializable_dict()
-        offer = self.offer_factory().serializable_dict()
+        offer = offer_factory().serializable_dict()
         assert PreferredPartnersMatchingAlgorithm._can_order_be_matched(
             bid=bid,
             offer=offer,
